port to flink 0.14
