+++ flink/v1.13/flink/src/test/java/org/apache/iceberg/flink/sink/TestFlinkManifest.java
@@ -224,18 +224,18 @@ public class TestFlinkManifest {
   }
 
   private DataFile writeDataFile(String filename, List<RowData> rows) throws IOException {
-    return SimpleDataUtil.writeDataFile(table, FileFormat.PARQUET, filename, appenderFactory, null, rows);
+    return SimpleDataUtil.writeFile(table.schema(), table.spec(), CONF,
+        tablePath, FileFormat.PARQUET.addExtension(filename), rows);
   }
 
   private DeleteFile writeEqDeleteFile(String filename, List<RowData> deletes) throws IOException {
-    return SimpleDataUtil
-        .writeEqDeleteFile(table, FileFormat.PARQUET, filename, appenderFactory, null, deletes);
+    return SimpleDataUtil.writeEqDeleteFile(table, FileFormat.PARQUET, tablePath, filename, appenderFactory, deletes);
   }
 
   private DeleteFile writePosDeleteFile(String filename, List<Pair<CharSequence, Long>> positions)
       throws IOException {
     return SimpleDataUtil
-        .writePosDeleteFile(table, FileFormat.PARQUET, filename, appenderFactory, null, positions);
+        .writePosDeleteFile(table, FileFormat.PARQUET, tablePath, filename, appenderFactory, positions);
   }
 
   private List<DataFile> generateDataFiles(int fileNum) throws IOException {
