+++ flink/v1.14/flink/src/main/java/org/apache/iceberg/flink/sink/IcebergFilesCommitter.java
@@ -21,6 +21,7 @@ package org.apache.iceberg.flink.sink;
 
 import java.io.IOException;
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
@@ -45,6 +46,7 @@ import org.apache.iceberg.RowDelta;
 import org.apache.iceberg.Snapshot;
 import org.apache.iceberg.SnapshotUpdate;
 import org.apache.iceberg.Table;
+import org.apache.iceberg.events.CreateSnapshotEvent;
 import org.apache.iceberg.flink.TableLoader;
 import org.apache.iceberg.io.WriteResult;
 import org.apache.iceberg.relocated.com.google.common.base.MoreObjects;
@@ -58,8 +60,8 @@ import org.apache.iceberg.util.PropertyUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-class IcebergFilesCommitter extends AbstractStreamOperator<Void>
-    implements OneInputStreamOperator<WriteResult, Void>, BoundedOneInput {
+class IcebergFilesCommitter extends AbstractStreamOperator<CommitResult>
+    implements OneInputStreamOperator<WriteResult, CommitResult>, BoundedOneInput {
 
   private static final long serialVersionUID = 1L;
   private static final long INITIAL_CHECKPOINT_ID = -1L;
@@ -276,6 +278,8 @@ class IcebergFilesCommitter extends AbstractStreamOperator<Void>
       }
 
       commitOperation(appendFiles, numFiles, 0, "append", newFlinkJobId, checkpointId);
+
+      emitCommitResult(appendFiles, pendingResults.values());
     } else {
       // To be compatible with iceberg format V2.
       for (Map.Entry<Long, WriteResult> e : pendingResults.entrySet()) {
@@ -299,6 +303,8 @@ class IcebergFilesCommitter extends AbstractStreamOperator<Void>
         Arrays.stream(result.deleteFiles()).forEach(rowDelta::addDeletes);
 
         commitOperation(rowDelta, numDataFiles, numDeleteFiles, "rowDelta", newFlinkJobId, e.getKey());
+
+        emitCommitResult(rowDelta, Collections.singleton(result));
       }
     }
   }
@@ -316,9 +322,24 @@ class IcebergFilesCommitter extends AbstractStreamOperator<Void>
     LOG.info("Committed in {} ms", duration);
   }
 
+  private void emitCommitResult(SnapshotUpdate<?> operation, Iterable<WriteResult> results) {
+    Preconditions.checkArgument(operation.updateEvent() instanceof CreateSnapshotEvent,
+        "Operation update event should be instance of CreateSnapshotEvent, but not %s", operation.updateEvent());
+
+    CreateSnapshotEvent event = (CreateSnapshotEvent) operation.updateEvent();
+    CommitResult commitResult = CommitResult.builder(event.sequenceNumber(), event.snapshotId())
+        .addAll(results)
+        .build();
+    emit(commitResult);
+  }
+
   @Override
   public void processElement(StreamRecord<WriteResult> element) {
     this.writeResultsOfCurrentCkpt.add(element.getValue());
+  }
+
+  private void emit(CommitResult result) {
+    output.collect(new StreamRecord<>(result));
   }
 
   @Override
