+++ flink/v1.14/flink/src/main/java/org/apache/iceberg/flink/sink/IcebergRewriteFilesCommitter.java
@@ -121,8 +121,8 @@ class IcebergRewriteFilesCommitter extends AbstractStreamOperator<Void>
   }
 
   @Override
-  public void dispose() throws Exception {
-    super.dispose();
+  public void close() throws Exception {
+    super.close();
     if (tableLoader != null) {
       tableLoader.close();
     }
