+++ flink/v1.14/flink/src/main/java/org/apache/iceberg/flink/sink/IcebergStreamRewriter.java
@@ -274,8 +274,8 @@ class IcebergStreamRewriter extends AbstractStreamOperator<RewriteResult>
   }
 
   @Override
-  public void dispose() throws Exception {
-    super.dispose();
+  public void close() throws Exception {
+    super.close();
     if (tableLoader != null) {
       tableLoader.close();
     }
